requirements.txt
selenium>=3.14
matplotlib>=3.1.2
pandas>=1.1.5
mplfinance>=0.12.4a0
pandas_datareader>=0.9.0
numpy>=1.19.5
pyzmq>=14.7.0
scikit-learn>=0.18
Pillow>=2.9.0
scipy>=0.19.0
beautifulsoup4>=4.9.3
sqlalchemy>=1.3.23
sqlalchemy-redshift>=0.8.2
psycopg2>=2.8.6

requests~=2.25.1
python-smsapi~=0.1.1